Move uploaders, clean up some remaining old code
