+++ src/client/rust/builders/object3_builder.rs
@@ -1,4 +1,3 @@
-use crate::artifact_uploader_3d::Type3d;
 use crate::builders::Geometry3Builder;
 use crate::builders::Transform3Builder;
 use artifacts_api_rust_proto::Object3;
@@ -29,11 +28,5 @@ impl Into<StructuredData> for &Object3Builder {
         let mut s = StructuredData::new();
         *s.mut_object3() = self.proto.clone();
         s
-    }
-}
-
-impl Type3d for Object3Builder {
-    fn convert_3d_to_raw(&self) -> StructuredData {
-        self.into()
     }
 }
