+++ src/client/rust/uploaders/mod.rs
@@ -1,0 +1,12 @@
+mod artifact_uploader_2d;
+mod artifact_uploader_3d;
+pub(crate) mod base_artifact_uploader;
+mod generic_artifact_uploader;
+mod run_stage_uploader;
+mod run_uploader;
+
+pub use artifact_uploader_2d::ArtifactUploader2d;
+pub use artifact_uploader_3d::ArtifactUploader3d;
+pub use generic_artifact_uploader::GenericArtifactUploader;
+pub use run_stage_uploader::RunStageUploader;
+pub use run_uploader::RunUploader;
