+++ src/client/rust/uploaders/run_stage_uploader.rs
@@ -1,7 +1,6 @@
-use crate::artifact_uploader_2d::ArtifactUploader2d;
-use crate::base_artifact_uploader::BaseArtifactUploader;
-use crate::builders::UserMetadataBuilder;
-use crate::generic_artifact_uploader::GenericArtifactUploader;
+
+use crate::uploaders::base_artifact_uploader::BaseArtifactUploader;
+
 use wasm_bindgen::prelude::*;
 
 #[wasm_bindgen]
@@ -11,6 +10,7 @@ pub struct RunStageUploader {
 
 #[wasm_bindgen]
 impl RunStageUploader {
+    /*
     pub fn child_uploader(&self, metadata: &UserMetadataBuilder) -> GenericArtifactUploader {
         self.base.child_uploader_old(metadata)
     }
@@ -18,4 +18,5 @@ impl RunStageUploader {
     pub fn child_uploader_2d(&self, metadata: &UserMetadataBuilder) -> ArtifactUploader2d {
         self.base.child_uploader_2d_old(metadata)
     }
+     */
 }
