+++ src/client/rust/util.rs
@@ -1,3 +1,4 @@
+use artifacts_api_rust_proto::ArtifactId;
 use protobuf::well_known_types::timestamp::Timestamp;
 use protobuf::Message;
 use std::error::Error;
@@ -30,6 +31,12 @@ impl Into<JsValue> for ClientError {
     fn into(self) -> JsValue {
         JsValue::from_str(&self.to_string())
     }
+}
+
+pub(crate) fn new_artifact_id() -> ArtifactId {
+    let mut id = ArtifactId::new();
+    id.uuid = Some(new_uuid_proto()).into();
+    id
 }
 
 pub(crate) fn encode_id_proto(msg: &impl Message) -> String {
